+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.2] - 2020-07-14
+
+### Changed
+
+- Optimized Text.pad
+
+### Added
+
+- Added rich.scope
+- Change log_locals to use scope.render_scope
+- Added title parameter to Columns
+
 ## [3.3.1] - 2020-07-13
 
 ### Added